requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
CacheControl==0.14.1
cachetools==5.5.0
celery==5.4.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
cryptography==44.0.0
Django==5.1.4
django-celery-beat==2.7.0
django-celery-results==2.5.1
django-ckeditor==6.7.2
django-cleanup==9.0.0
django-cors-headers==4.6.0
django-debug-toolbar==4.4.6
django-filter==24.3
django-js-asset==3.0.1
django-phonenumber-field==8.0.0
django-resized==1.0.3
django-rest-registration==0.9.0
django-timezone-field==7.0
djangorestframework==3.15.2
drf-writable-nested==0.7.1
drf-yasg==1.21.8
fcm-django==2.2.1
firebase-admin==6.6.0
google-api-core==2.24.0
google-api-python-client==2.156.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.19.0
google-cloud-storage==2.19.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
grpcio==1.68.1
grpcio-status==1.68.1
httplib2==0.22.0
idna==3.10
inflection==0.5.1
kombu==5.4.2
msgpack==1.1.0
packaging==24.2
phonenumbers==8.13.52
pillow==11.0.0
prompt_toolkit==3.0.48
proto-plus==1.25.0
protobuf==5.29.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.10.1
pyparsing==3.2.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
rsa==4.9
six==1.17.0
sqlparse==0.5.3
swapper==1.4.0
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
xmltodict==0.14.2
